feat: Alternative implementation of JSON logic
Attempt to implement the JSON logic functionality based on three data types: `Expr`, `Rule` and `RuleSet`. Any valid JSON value and variables can be converted to an `Expr`. For now, `Rule` support various comparison operators that combine two `Expr`. A `RuleSet` defines either any of all of the rules must be satisfied. To create more complex logic, one can create a sub ruleset and attach it to a `RuleSet` as a rule.