feat: Theme installation util
Develop a CLI utility that processes a JSON theme file and generates a corresponding CSS file based on the specified parameters.

To use it, go to /misc/themeinstall, then:
cargo run -- --input [path to the input JSON file] --output [path to the output directory]
The tool will automatically name the output css using the `slug` field.