feat(moss-typebridge): Implement basic struct export functionality
Create a new attribute macro `#[type_bridge]`, which currently supports generating TypeScript typing for simple structs. The struct will first be parsed into an abstract representation `StructDefinition`, which decouples the parsing logic from exporting logic. This means that in the future we can easily add functionalities to export to new languages.

To test it, go to `/crates/moss-typebridge` and run `cargo run`. It will currently print the exported TypeScript type in the console; in the future we will add an attribute flag to indicate the export file path.