perf: Change cargo feature resolver version to 1
The version 2 of cargo feature resolver is more conservative than version 1: it will not automatically merge feature sets for dependencies that are used both as build dependencies and normal dependencies, which results in packages like `tauri-utils` being compiled into two targets because of different feature sets requirement. While theoretically a more liberal approach to feature merging could lead to problems when there are incompatible features, this shouldn't be a huge issue for our project since we are primarily aimed at desktop development at this stage.
https://doc.rust-lang.org/edition-guide/rust-2021/default-cargo-resolver.html