perf: Disable crate type specification for desktop app lib
According to tauri documentation, we only need to specify the desktop lib to multiple formats when we need to produce libraries for mobile app; this means that for now this won't be necessary. Furthermore, it looks like dynamic libraries must be recompiled every time, explaining why we sometimes have recompilation when nothing has been changed.