feat(moss-cache): Abstracting and modularizing cache backend
To decouple our final cache from the exact implementation of the cache we use (right now we are using `moka`'s sync cache), I have abstracted the backend logic into a `CacheBackend` trait, which hides its own way of implementing the ttl and expiry mechanism away from the frontend. At the end we expose a generic struct that allow us to switch the backend while preserving the frontend API with the `CacheFrontend` trait, although we need to implement the frontend method for different backend modules.