perf(MOSS-335): Dependency cleanup
While cargo-udeps doesn't reveal much, a manual review of the crate's and workspace's toml reveals that there are actually a decent number of unused dependencies after the refactoring. I also did a few profilings and auditing with `cargo tree` but have not found out any particular dependencies that need to be removed.