feat(MOSSMVP-394): WASM plugin proof of concept
A basic demo of compiling a WASM plugin, incorporating it into the tauri app, and two way function calling between the plugin and the function host. This is a very preliminary demonstration, so there's a lot more to be worked on, including a separate WASM host to handle WASM related logic, better design of public API for the plugins, integration with our extension mechanism, and management of plugin states.

Here's the step for creating and compiling the wasm plugin:

1. The `view/desktop/bin/src/wasm` stores the wasm library to be compiled. The `Cargo.toml` should specify crate-type as `cdylib`, and it uses wit-bindgen to generate Rust bindings for the WIT definition.
2. `/wasm/wit` stores the WIT files, which defines the contracts between components. The exact implementation of these interfaces can be done in different languages that can compile into WASM, making it possible for the users to create plugins using languages of their choice in the future. (https://component-model.bytecodealliance.org/design/wit.html)
3. `wasm/src/lib.rs` uses wit_bindgen to parse the WIT and allow you to provide implementation of the interface using the Rust trait system. For the demo plugin, we define an `execute` function to be exported by the plugin that simply calls the `open_new_window` function imported from the host.
4. To compile the wasm plugin, first we need to `rustup add target wasm32-wasip2`, and use `cargo build --target wasm32-wasip2` to compile the wasm file.
5. We will then need to move the wasm file from the target folder so that the main app can access it. However, I am still struggling with pathing resolution in Rust, so I just put it at both the `desktop/bin` and the `target` folders.
6. In the `view/desktop/bin/src/lib.rs`, we initialized a wasm environment, imported the wasm byte-code, compile it to a component, and provide the host definition of the `open_new_window` function. Then, we will call the plugin's `execute` function once the `Ctrl-N` shortcut is triggered.
(Warning: I have not figured out how to disable the trace logging for wasmtime and cranelift yet, so you will be bombarded with tracing logs for 1-2 minutes after pressing Ctrl-N. It's actually the compilation process. In the future, we will optimize this process by only compiling the wasm code once at plugin installation/update, and serialize the compiled plugin for normal app execution.)
7. If everything goes well, after 1-2 minutes, a new window will be opened.